backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
backports.asyncio.runner==1.2.0
banks==2.3.0
beautifulsoup4==4.14.3
Bottleneck @ file:///home/task_176193794808509/conda-bld/bottleneck_1761937988290/work
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
ecdsa==0.19.1
et_xmlfile @ file:///croot/et_xmlfile_1756462486635/work
exceptiongroup==1.3.1
fastapi==0.135.1
filelock==3.25.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.1.0
greenlet==3.3.1
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.18
llama-index-cli==0.5.6
llama-index-core==0.14.18
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-ollama==0.9.0
llama-index-embeddings-openai==0.6.0
llama-index-graph-stores-neo4j==0.7.0
llama-index-indices-managed-llama-cloud==0.10.0
llama-index-instrumentation==0.5.0
llama-index-llms-ollama==0.10.0
llama-index-llms-openai==0.7.2
llama-index-readers-file==0.5.6
llama-index-readers-llama-parse==0.6.0
llama-index-workflows==2.17.0
llama-parse==0.5.20
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mkl-service==2.5.2
mkl_fft @ file:///home/task_176159252575002/conda-bld/mkl_fft_1761592901585/work
mkl_random @ file:///home/task_176159293813909/conda-bld/mkl_random_1761592952512/work
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
neo4j==5.28.3
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.3
numexpr @ file:///home/task_176216559424661/conda-bld/numexpr_1762165604917/work
numpy @ file:///home/task_176157872663408/conda-bld/numpy_and_numpy_base_1761579342634/work/dist/numpy-2.2.5-cp310-cp310-linux_x86_64.whl#sha256=fb5b85b589f86e6404be6a37347b98be70b32c4ae837bb26e1180e9f9a7fb718
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
ollama==0.6.1
openai==2.29.0
openpyxl==3.1.5
packaging==26.0
pandas @ file:///home/task_176233228161209/conda-bld/pandas_1762332325727/work/dist/pandas-2.3.3-cp310-cp310-linux_x86_64.whl#sha256=cd4241f338dbf9922b2bc5a62ff6267574e607e9e9ea8b5efbba07b037c96f8f
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
propcache==0.4.1
pyasn1==0.6.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.26.7
pymupdf4llm==0.2.9
pypdf==6.9.1
pypdfium2==5.6.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytesseract==0.3.13
pytz @ file:///croot/pytz_1752135852232/work
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.3
shellingham==1.5.4
six @ file:///croot/six_1744271502820/work
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
striprtf==0.0.26
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tinytag==2.2.1
tokenizers==0.22.2
tomli==2.4.0
torch==2.10.0
tqdm==4.67.1
transformers==5.2.0
triton==3.6.0
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata @ file:///croot/python-tzdata_1746123641790/work
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==1.17.3
yarl==1.22.0
